+++ Temp_Spring/src/main/java/com/example/temp_spring/domain/user/TempUser.java
@@ -24,5 +24,5 @@ public class TempUser {
     private String password;
     private String email;
     private UserRole role;
-    private Long farmId;
+    private int farmId;
 }
