+++ Temp_Spring/src/main/java/com/example/temp_spring/domain/user/User.java
@@ -29,5 +29,5 @@ public class User {
     private String email;
     private UserRole role;
     private String salt;
-    private Long farmId;
+    private int farmId;
 }
